requirements.txt
amqp
beautifulsoup4
boto3
celery[brotli]
django
django-storages
django_neomodel
flower
ipython
lxml
pendulum
psycopg2-binary
redis
requests
sentry-sdk
tqdm
